[refactor] 랭킹 조회 대상 및 조회 조건 enum 클래스 디렉토리 위치 변경 (KakaoFunding#183)
